+++ server/src/main/java/org/elasticsearch/ExceptionsHelper.java
